chore: split TaskGroupInner into separate file
